requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
pillow>=10.0.0
bitsandbytes>=0.41.0
accelerate>=0.23.0
